Remove post-build command for AsmFuncs.vasm
Do not delete AsmFuncs.vasm after the build finishes. This adds unnecessary generate-compile-build-remove chain which does not have any benefit.

Once this custom command is removed, the AsmFuncs.vasm will be regenerated every time the source file (AsmFuncs.a) changes or its timestamp updates. This is the correct way of doing this.